+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.15.3
+* Update CPU architectures based on Myrsloik's "essay": https://github.com/vapoursynth/vapoursynth/issues/1185#issuecomment-4235066569
+* This means we now target Haswell instead of the generic `x86_64_v3` target, as well as `Zen 4` albeit with SSE4a
+turned off (which was never supported by Intel).
+
 ## 0.15.2
 * Actually build Python modules with AVX512 turned on for x86_64_v4 targets